server/requirements.txt
Django==1.9.2
FeinCMS==1.11.4
Pillow==3.1.1
PyJWT==1.4.0
PyYAML==3.11
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.22
celery==3.1.20
celery-with-redis==3.0
certifi==2015.11.20.1
django-angular==0.7.16
django-cacheback==1.0
django-celery==3.1.17
django-cors-headers==1.1.0
-e git+https://github.com/toastdriven/django-haystack.git@12f54227903acdfcc9f5dc1ae918cec9396b2900#egg=django_haystack-master
django-mptt==0.8.0
django-redis-cache==1.6.5
django-rest-swagger==0.3.5
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
drf-haystack==1.5.6
elasticsearch==1.6.0
feedparser==5.2.1
gevent==1.0.2
greenlet==0.4.9
haystack==0.33
kombu==3.0.33
pefile==1.2.10-114
pyelasticsearch==1.4
python-dateutil==2.4.2
python-memcached==1.57
python-ptrace==0.8.1
pytz==2015.7
redis==2.10.5
requests==2.9.1
simplejson==3.8.2
six==1.10.0
urllib3==1.14
wheel==0.24.0
wsgiref==0.1.2
